feat: init android

+++ rust-libs/ease-client/src/modules/storage/typ.rs
@@ -5,7 +5,7 @@ use crate::define_id;
 
 define_id!(StorageId);
 
-#[derive(FromPrimitive, ToPrimitive, Serialize, Clone, Debug, PartialEq, Eq)]
+#[derive(FromPrimitive, ToPrimitive, Serialize, Clone, Debug, PartialEq, Eq, uniffi::Enum)]
 pub enum StorageType {
     Local,
     Webdav,
@@ -18,7 +18,7 @@ impl Default for StorageType {
     }
 }
 
-#[derive(Debug, Clone, Serialize, PartialEq, Eq)]
+#[derive(Debug, Clone, Serialize, PartialEq, Eq, uniffi::Enum)]
 pub enum StorageEntryType {
     Folder,
     Music,
@@ -27,7 +27,7 @@ pub enum StorageEntryType {
     Other,
 }
 
-#[derive(Debug, Clone, Copy, Serialize, PartialEq, Eq)]
+#[derive(Debug, Clone, Copy, Serialize, PartialEq, Eq, uniffi::Enum)]
 pub enum CurrentStorageImportType {
     Musics,
     EditPlaylistCover,
@@ -36,7 +36,7 @@ pub enum CurrentStorageImportType {
     CurrentMusicLyrics,
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, uniffi::Enum)]
 pub enum StorageConnectionTestResult {
     None,
     Testing,
@@ -65,7 +65,7 @@ pub struct StorageInfo {
     pub typ: StorageType,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Default)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Default, uniffi::Record)]
 pub struct ArgUpsertStorage {
     pub id: Option<StorageId>,
     pub addr: String,
@@ -76,7 +76,7 @@ pub struct ArgUpsertStorage {
     pub typ: StorageType,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, uniffi::Enum)]
 pub enum CurrentStorageStateType {
     Loading,
     OK,
@@ -92,7 +92,7 @@ impl Default for CurrentStorageStateType {
     }
 }
 
-#[derive(Debug, Clone, Serialize)]
+#[derive(Debug, Clone, Serialize, uniffi::Record)]
 pub struct VStorageListItem {
     pub storage_id: StorageId,
     pub name: String,
@@ -100,12 +100,12 @@ pub struct VStorageListItem {
     pub typ: StorageType,
 }
 
-#[derive(Debug, Clone, Default, Serialize)]
+#[derive(Debug, Clone, Default, Serialize, uniffi::Record)]
 pub struct VStorageListState {
     pub items: Vec<VStorageListItem>,
 }
 
-#[derive(Debug, Clone, Serialize)]
+#[derive(Debug, Clone, Serialize, uniffi::Record)]
 pub struct VCurrentStorageEntry {
     pub path: String,
     pub name: String,
@@ -115,13 +115,13 @@ pub struct VCurrentStorageEntry {
     pub entry_typ: StorageEntryType,
 }
 
-#[derive(Debug, Clone, Serialize)]
+#[derive(Debug, Clone, Serialize, uniffi::Record)]
 pub struct VSplitPathItem {
     pub path: String,
     pub name: String,
 }
 
-#[derive(Debug, Clone, Serialize)]
+#[derive(Debug, Clone, Serialize, uniffi::Record)]
 pub struct VCurrentStorageEntriesStateStorageItem {
     pub id: StorageId,
     pub name: String,
@@ -130,7 +130,7 @@ pub struct VCurrentStorageEntriesStateStorageItem {
     pub is_local: bool,
 }
 
-#[derive(Debug, Clone, Serialize)]
+#[derive(Debug, Clone, Serialize, uniffi::Record)]
 pub struct VCurrentStorageEntriesState {
     pub import_type: CurrentStorageImportType,
     pub state_type: CurrentStorageStateType,
@@ -143,13 +143,13 @@ pub struct VCurrentStorageEntriesState {
     pub disabled_toggle_all: bool,
 }
 
-#[derive(Debug, Clone, Serialize)]
+#[derive(Debug, Clone, Serialize, uniffi::Record)]
 pub struct VEditStorageState {
     pub is_created: bool,
     pub title: String,
     pub info: ArgUpsertStorage,
     pub test: StorageConnectionTestResult,
-    pub music_count: usize,
-    pub playlist_count: usize,
+    pub music_count: u32,
+    pub playlist_count: u32,
     pub update_signal: u16,
 }
